+++ src/frontend/src/pages/EvmToEvm.js
@@ -120,7 +120,7 @@ const EvmToEvm = ({
   }, [chains]);
 
   return (
-    <div className="flex-grow max-w-3xl mx-auto p-4">
+    <div className="w-full p-4">
       <h1 className="text-2xl font-bold text-center mb-6">EVM to EVM Swap</h1>
 
       {/* Step 1 */}
