+++ src/frontend/src/pages/EvmToLightning.js
@@ -130,7 +130,7 @@ const EvmToLightning = ({
   }, [netId]);
 
   return (
-    <div className="flex-grow max-w-3xl mx-auto p-4">
+    <div className="w-full p-4">
       <h1 className="text-2xl font-bold text-center mb-6">EVM to Lightning Swap</h1>
 
       {/* Step 1 */}
