+++ src/frontend/src/pages/LightningToEvm.js
@@ -81,7 +81,7 @@ const LightningToEvm = ({ chains, coinbase }) => {
   }, [chains]);
 
   return (
-    <div className="flex-grow max-w-3xl mx-auto p-4">
+    <div className="w-full p-4">
       <h1 className="text-2xl font-bold text-center mb-6">Lightning to EVM Swap</h1>
 
       {/* Step 1 */}
