+++ src/frontend/src/pages/Petitions.js
@@ -5,6 +5,11 @@ import { ethers } from 'ethers';
 import ERC20ABI from '../../assets/contracts/abis/erc20Abi.json';
 import { main } from "../../../declarations/main";
 
+
+import CreatePetition from "../components/petitions/CreatePetition";
+import SolvePetitions from "../components/petitions/SolvePetitions";
+
+
 const Petitions = ({
   coinbase,
   netId,
@@ -104,6 +109,7 @@ const Petitions = ({
       }
       const signer = await provider.getSigner();
       if (solve && !(Number(petitionToSolve.current.wantedChain) === Number(netId))) {
+        setProcessing(false);
         alert("Wrong network");
         return;
       }
@@ -168,7 +174,7 @@ const Petitions = ({
   }, [coinbase]);
 
   return (
-    <div className="flex-grow max-w-3xl mx-auto p-4">
+    <div className="w-full p-4">
       <h1 className="text-2xl font-bold text-center mb-6">Petitions</h1>
 
       <div className="flex space-x-4 mb-6">
@@ -192,110 +198,28 @@ const Petitions = ({
 
       {
         !solve ?
-          <div className="mb-6">
-            <div className="mb-4">
-              <p>Step 1: Select recipient and EVM compatible chain</p>
-              <label className="block mb-2">EVM Recipient Address</label>
-              <input
-                className="w-full p-2 border border-gray-300 rounded mb-4"
-                value={evm_address}
-                onChange={(ev) => setEvmAddr(ev.target.value)}
-                placeholder="Enter EVM address"
-              />
-              <label className="block mb-2">Select Destiny Chain</label>
-              <select
-                className="w-full p-2 border border-gray-300 rounded mb-4"
-                onChange={(ev) => setChain(ev.target.value)}
-              >
-                {
-                  chains.map(item => {
-                    const filteredRpc = item.rpc.filter(rpcUrl => !rpcUrl.includes("${INFURA_API_KEY}"));
-                    if (filteredRpc.length > 0) {
-                      return (
-                        <option value={JSON.stringify({
-                          rpc: filteredRpc[0].toString(),
-                          chainId: item.chainId,
-                          name: item.name
-                        })}>{item.name}</option>
-                      );
-                    } else {
-                      return null;
-                    }
-                  })
-                }
-              </select>
-              {
-                chain &&
-                <>
-                  <p>Bridging to {JSON.parse(chain).name}</p>
-                  <p>ChainId {JSON.parse(chain).chainId}</p>
-                </>
-              }
-            </div>
-            <div className="mb-4">
-              <p>Step 2: Send token to 0x{canisterAddr}</p>
-              <label className="block mb-2">Amount in satoshis</label>
-              <input
-                className="w-full p-2 border border-gray-300 rounded mb-4"
-                value={amount}
-                onChange={(ev) => setAmount(ev.target.value)}
-                placeholder="Satoshis"
-              />
-              {
-                !coinbase ?
-                  <button className="w-full p-2 mt-2 bg-blue-500 text-white rounded" onClick={loadWeb3Modal}>Connect Wallet</button> :
-                  !processing ?
-                    <button className="w-full p-2 mt-2 bg-blue-500 text-white rounded" onClick={() => { sendToken(solve); }}>Send token</button> :
-                    <button className="w-full p-2 mt-2 bg-gray-400 text-white rounded" disabled>Wait current process</button>
-              }
-            </div>
-            {
-              coinbase && netId &&
-              <div className="mb-4">
-                <p>Sending from chainId {netId.toString()}</p>
-              </div>
-            }
-          </div> :
-          <div className="mb-6">
-            <div className="mb-4">
-              <h2>Petitions</h2>
-              {
-                petitions.map(item => {
-                  if (Number(netId) !== Number(item.wantedChain)) return;
-                  if (item.sendingChain === "0" || item.wantedChain === "0") return;
-                  return (
-                    <div key={item.proofTxId} className="mb-4">
-                      <p>From chain: {item.sendingChain}</p>
-                      <p>To chain: {item.wantedChain}</p>
-                      {
-                        item.sendingChain === "0x1f" ?
-                          <p>Amount: {(Number(item.transaction.value) / 10 ** 10)?.toString()} satoshis of rbtc</p> :
-                          <p>Amount: {(Number(`0x${item.transaction.data.slice(74).replace(/^0+/, '')}`) / 10 ** 10).toString()} satoshis of wbtc</p>
-                      }
-                      <p>Reward: {item.reward}</p>
-                      {
-                        petitionToSolve.current &&
-                        (
-                          JSON.stringify(petitionToSolve.current) === JSON.stringify(item) &&
-                          <p><b>Petition Selected</b></p>
-                        )
-                      }
-                      <button className="w-full p-2 mt-2 bg-blue-500 text-white rounded" onClick={async () => {
-                        petitionToSolve.current = item;
-                        sendToken(solve);
-                      }}>Initiate petition solving</button>
-                      {
-                        !petitionToSolve.current &&
-                        <button className="w-full p-2 mt-2 bg-blue-500 text-white rounded" onClick={async () => {
-                          petitionToSolve.current = item;
-                        }}>Select Petition</button>
-                      }
-                    </div>
-                  );
-                })
-              }
-            </div>
-          </div>
+        <CreatePetition
+          canisterAddr={canisterAddr}
+          chain={chain}
+          chains={chains}
+          setChain={setChain}
+          setEvmAddr={setEvmAddr}
+          setAmount={setAmount}
+          sendToken={sendToken}
+          netId={netId}
+          evm_address={evm_address}
+          amount={amount}
+          coinbase={coinbase}
+          processing={processing}
+          loadWeb3Modal={loadWeb3Modal}
+        /> :
+        <SolvePetitions
+          petitions={petitions}
+          petitionToSolve={petitionToSolve}
+          netId={netId}
+          sendToken={sendToken}
+          solve={solve}
+        />
       }
 
       <div className="mb-6">
